Move cart into a separate app
Closes #17
