+++ cart/admin.py
@@ -1,0 +1,6 @@
+from django.contrib import admin
+
+from .models import Cart, CartItem
+
+admin.site.register(Cart)
+admin.site.register(CartItem)
