+++ clothes/apps.py
@@ -4,6 +4,3 @@ from django.apps import AppConfig
 class ClothesConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
     name = "clothes"
-
-    def ready(self):
-        from . import signals
