+++ clothes/serializers.py
@@ -1,6 +1,6 @@
 from rest_framework import serializers
 
-from .models import CartItem, Category, Cloth, Image
+from .models import Category, Cloth, Image
 
 
 class ImageSerializer(serializers.ModelSerializer):
@@ -85,24 +85,3 @@ class ClothCreateSerializer(serializers.ModelSerializer):
             "id",
             "cover_img_url",
         )
-
-
-class CartItemSerializer(serializers.ModelSerializer):
-    cloth = ClothSerializer()
-
-    class Meta:
-        model = CartItem
-        fields = (
-            "id",
-            "cloth",
-        )
-
-
-class CartItemCreateSerializer(serializers.ModelSerializer):
-    class Meta:
-        model = CartItem
-        fields = (
-            "cart",
-            "cloth",
-        )
-        read_only_fields = ("cart",)
