+++ clothes/urls.py
@@ -9,6 +9,4 @@ urlpatterns = [
     path("colors", views.ColorList.as_view()),
     path("sizes", views.SizeList.as_view()),
     path("price-range", views.SellPriceRangeView.as_view()),
-    path("cart", views.CartItemsList.as_view()),
-    path("cart-items/<int:pk>", views.CartItemsView.as_view()),
 ]
